requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.6.0
triton-windows==3.2.0.post21
numpy==2.4.0
matplotlib==3.10.8
termcolor==3.2.0
rich==14.2.0
pandas==3.0.2